--- SilverPlatter.Server/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "dotnet": "^1.1.4"
-  }
-}
