+++ silverplatter.client/src/Pages/Browse.tsx
@@ -1,4 +1,4 @@
-
+import './css/Browse.css'
 
 function Browse() {
     return (
