+++ silverplatter.client/src/Pages/LandingPage.tsx
@@ -1,41 +1,18 @@
-
+import './css/LandingPage.css'
 
 function LandingPage() {
 
 
     return (
-        <div style={{
-            width: "100vw",
-            display: "flex",
-            flexDirection: "column",
-            position: "relative",
-        }}>
-            <div id="Content" style={{
-                backgroundColor:"rgba(0, 0, 0, 0.5)",
-                height: "100vh",
-                order: 1,
-                zIndex: 2,
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "center"
-            }}>
+        <div id="LandingPage">
+            <div id="Content">
                 <img 
                     src="src/assets/claim-your-i-was-here-button-v0-9tlo8368wdkf1.webp" 
-                    style={{
-                        maxWidth: "500px",
-                        height: "auto"
-                    }}
                 />
             </div>
 
-            <div id="Background" style={{
-                order: 2, 
-                zIndex: 1,
-                marginTop: "-100vh"
-            }}>
-                <video autoPlay muted loop style={{
-                    width: "100%",
-                }}> 
+            <div id="Background">
+                <video autoPlay muted loop> 
                     <source src="src/assets/LandingBackground.mp4" type="video/mp4"/>
                 </video>
             </div>
