+++ silverplatter.client/src/Pages/css/Header.css
@@ -1,0 +1,20 @@
+
+#Header {
+    width: 100vw;
+    height: 5rem;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    background-color: darkslategrey;
+    outline: solid 2px black;
+}
+
+#Menu {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    justify-content: center;
+    gap: 5%;
+}
