+++ silverplatter.client/src/Types/BookableTable.ts
@@ -1,0 +1,9 @@
+
+
+export interface BookableTable {
+    Id: number;
+    Name: string;
+    Description: string;
+    Places: number;
+    RestaurantId: number;
+}
