+++ silverplatter.client/src/Types/MenuItem.ts
@@ -1,0 +1,8 @@
+
+export interface MenuItem {
+    Id: number;
+    Name: string;
+    Description: string;
+    RestaurantId: number;
+}
+
