+++ silverplatter.client/src/Types/Restaurant.ts
@@ -1,0 +1,8 @@
+
+
+export interface Restaurant {
+    Id : number;
+    name: string;
+    Description: string;
+    Address: string;
+}
